rename methods + namespaces
