+++ pyconfigurableml/logging.py
@@ -1,13 +1,10 @@
 import logging
-from pyconfigurableml._decorators import pass_decorator
+from pyconfigurableml._core import run_with_specified_config, T
 from typeguard import typechecked
 from typing import Dict, TypeVar, Union
 
 
-T = TypeVar('T')
-
-
-@pass_decorator(__name__)
+@run_with_specified_config(__name__)
 @typechecked
 def set_logger_levels(config: T, inner_config: Dict[str, str]) -> T:
     '''
